+++ src/main/java/org/folio/search/configuration/kafka/KafkaConfiguration.java
@@ -50,8 +50,8 @@ public abstract class KafkaConfiguration {
   }
 
   public enum SearchTopic implements FolioKafkaTopic {
-    CONSORTIUM_INSTANCE("search.consortium.instance"),
-    REINDEX_RANGE_INDEX("search.reindex.range-index");
+    REINDEX_RANGE_INDEX("search.reindex.range-index"),
+    INDEX_SUB_RESOURCE("search.index.sub-resource");
 
     private final String topicName;